+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
+### Added
+- Implemented `server/listiso` and `server/mountiso` endpoints.
+### Changed
+- Add `ISOFile` attribute in `ServerDetails`.
 
 ## [8.3.1] - 2024-09-19
 ### Changed